backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
SQLAlchemy==2.0.23
dashscope
python-dotenv==1.0.0
requests==2.31.0